:sparkles:feat: 쿼리파라미터 sizeId, colorId 타입을 int -> Long 으로 변환하면서 조회 기능을 querydsl사용해서 구현, 소프트 딜리트 된 레코드는 조회되지 않도록 수정

+++ src/main/java/colorful/starbucks/product/application/ProductDetailService.java
@@ -13,7 +13,7 @@ public interface ProductDetailService {
     ProductOptionListResponseDto getProductOptionList(String productCode);
 
     ProductDetailCodeAndQuantityResponseDto getProductDetailWithOptions(
-            String productCode, int sizeId, int colorId
+            String productCode, Long sizeId, Long colorId
     );
 
     ProductDetailResponseDto getProductDetail(String productDetailCode);
